@@ -5,7 +5,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.1"
+        "php": "^7.4|^8.0"
     },
     "require-dev": {
         "malukenho/docheader": "~0.1",
