@@ -23,63 +23,64 @@ use Magento\Framework\Profiler;
 use Magento\Store\Model\Store;
 use Zend_Db_Expr;
 use Zend_Db_Select;
+use Zend_Db_Statement_Exception;
 
 class EavIterator implements IteratorAggregate
 {
     /**
      * @var int
      */
-    protected $batchSize;
+    protected int $batchSize;
 
     /**
      * @var string
      */
-    protected $entityCode;
+    protected string $entityCode;
 
     /**
      * @var EavConfig
      */
-    protected $eavConfig;
+    protected EavConfig $eavConfig;
 
     /**
      * @var AbstractAttribute[]
      */
-    protected $attributes = [];
+    protected array $attributes = [];
 
     /**
      * @var AbstractAttribute[]
      */
-    protected $attributesByCode = [];
+    protected array $attributesByCode = [];
 
     /**
      * @var Store
      */
-    protected $store;
+    protected Store $store;
 
     /**
      * @var int[]
      */
-    protected $entityIds = [];
+    protected array $entityIds = [];
 
     /**
      * @var Helper
      */
-    protected $helper;
+    protected Helper $helper;
 
     /**
      * @var DbContext
      */
-    protected $dbContext;
+    protected DbContext $dbContext;
 
     /**
      * @var \ArrayIterator[]
      */
-    protected $entitySet;
+    protected array $entitySet;
 
     /**
      * @var array
      */
-    protected $eavSelectOrder = [
+    protected array $eavSelectOrder = [
         'entity_id',
         'store_id',
     ];
@@ -87,7 +88,7 @@ class EavIterator implements IteratorAggregate
     /**
      * @var array
      */
-    protected $entityBatchOrder = [];
+    protected array $entityBatchOrder = [];
 
     /**
      * EavIterator constructor.
@@ -174,9 +175,9 @@ class EavIterator implements IteratorAggregate
     /**
      * @param MysqlStatement $stmt
      * @return \Generator
-     * @throws \Zend_Db_Statement_Exception
+     * @throws Zend_Db_Statement_Exception
      */
-    protected function loopUnionRows(MysqlStatement $stmt)
+    protected function loopUnionRows(MysqlStatement $stmt): \Generator
     {
         $entity = ['entity_id' => null];
         while ($row = $stmt->fetch()) {
