@@ -26,37 +26,37 @@ class Products implements WriterInterface
     /**
      * @var Config
      */
-    protected $config;
+    protected Config $config;
 
     /**
      * @var Iterator
      */
-    protected $iterator;
+    protected Iterator $iterator;
 
     /**
      * @var StoreManager
      */
-    protected $storeManager;
+    protected StoreManager $storeManager;
 
     /**
      * @var Helper
      */
-    protected $helper;
+    protected Helper $helper;
 
     /**
      * @var Logger
      */
-    protected $log;
+    protected Logger $log;
 
     /**
      * @var EavConfig
      */
-    protected $eavConfig;
+    protected EavConfig $eavConfig;
 
     /**
      * @var array
      */
-    protected $attributeOptionMap = [];
+    protected array $attributeOptionMap = [];
 
     /**
      * Products constructor.
@@ -125,7 +125,7 @@ class Products implements WriterInterface
         $this->iterator->setEntityIds([]);
 
         foreach ($this->iterator as $index => $data) {
-            $this->writeProduct($xml, $store->getId(), $data);
+            $this->writeProduct($xml, (int) $store->getId(), $data);
             // Flush every so often
             if ($index % 100 === 0) {
                 $writer->flush();
@@ -142,7 +142,7 @@ class Products implements WriterInterface
      * @param int $storeId
      * @param array $data
      */
-    protected function writeProduct(XmlWriter $xml, $storeId, array $data): void
+    protected function writeProduct(XmlWriter $xml, int $storeId, array $data): void
     {
         $xml->startElement('item');
 
@@ -188,8 +188,8 @@ class Products implements WriterInterface
      */
     public function writeAttribute(
         XmlWriter $xml,
-        $storeId,
-        $name,
+        int $storeId,
+        string $name,
         $attributeValue
     ): void
     {
@@ -214,7 +214,7 @@ class Products implements WriterInterface
      * @param AbstractAttribute $attribute
      * @return string[]
      */
-    protected function getAttributeOptionMap($storeId, AbstractAttribute $attribute): array
+    protected function getAttributeOptionMap(int $storeId, AbstractAttribute $attribute): array
     {
         $attributeKey = $storeId . '-' . $attribute->getId();
         if (!isset($this->attributeOptionMap[$attributeKey])) {
@@ -358,7 +358,6 @@ class Products implements WriterInterface
         }
 
         $result = [];
-        /** @var string $attributeValue */
         foreach ($values as $attributeValue) {
             $map = $this->getAttributeOptionMap($storeId, $attribute);
             $result[] = $map[$attributeValue] ?? null;
