@@ -20,5 +20,5 @@ class ExportEntityConfigurable extends CompositeExportEntity
     /**
      * @var bool
      */
-    protected $isStockCombined;
+    protected bool $isStockCombined;
 }
